Upload of Baccarat Hand Generator
Able to properly simulate shoes of baccarat. The cut from the front is randomized based on a card drawn from the shoe. The cut from the back can be any integer, currently hard coded as 14. Card memory is present. Every card value is recorded in the results tuple. This will allow card counting systems to be tested. 

Next will add "banker win", "player win", or "tie", as one more data point to record for the analysis section of the simulator.